+++ tests/Earth/Period/HourTest.php
@@ -94,7 +94,7 @@ class HourTest extends TestCase
         );
     }
 
-    public function cases()
+    public static function cases()
     {
         return [
             [20, 0, 20],