+++ tests/Earth/Period/MinuteTest.php
@@ -98,7 +98,7 @@ class MinuteTest extends TestCase
         );
     }
 
-    public function cases()
+    public static function cases()
     {
         return [
             [20, 0, 0, 20],