+++ tests/Earth/Period/SecondTest.php
@@ -99,7 +99,7 @@ class SecondTest extends TestCase
         );
     }
 
-    public function cases()
+    public static function cases()
     {
         return [
             [20, 0, 0, 0, 20],