+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Require `innmind/black-box` `5`
+
+### Removed
+
+- Support for PHP `8.1`
+
 ## 3.2.0 - 2023-02-11
 
 ### Added